[Feature #165145698] Added first proverb
