+++ app/src/main/res/drawable/gradient.xml
@@ -1,9 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
     <gradient
-        android:startColor="@color/gradient_one_start"
+        android:angle="45"
         android:endColor="@color/gradient_one_end"
-        android:angle="45"/>
+        android:startColor="@color/gradient_one_start" />
 </shape>
