+++ app/src/main/res/drawable/gradient_four.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
+    xmlns:android="http://schemas.android.com/apk/res/android">
     <gradient
         android:startColor="@color/gradient_five_start"
         android:centerColor="@color/gradient_five_center"
