+++ app/src/main/res/drawable/gradient_six.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android">
+    <gradient
+        android:startColor="@color/gradient_seven_start"
+        android:endColor="@color/gradient_seven_end"
+        android:angle="45"/>
+</shape>
