+++ app/src/main/res/drawable/gradient_three.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
+    xmlns:android="http://schemas.android.com/apk/res/android">
     <gradient
         android:startColor="@color/gradient_three_start"
         android:endColor="@color/gradient_three_end"
